Rename update_ranges_impl
We use "uv run datacube-ows-update"
to get the entire environment now,
so delete update_ranges.py and
rename update_ranges_impl.py
to update_ranges.py.
