+++ integration_tests/test_version.py
@@ -9,7 +9,7 @@ https://click.palletsprojects.com/en/7.x/testing/
 """
 
 from datacube_ows import __version__
-from datacube_ows.update_ranges_impl import main
+from datacube_ows.update_ranges import main
 
 
 def test_updates_ranges_version(runner) -> None:
