+++ pyproject.toml
@@ -97,7 +97,7 @@ all = ["datacube_ows[dev,ops,test]"]
 [project.scripts]
 datacube-ows = "datacube_ows.wsgi:main"
 datacube-ows-cfg = "datacube_ows.cfg_parser_impl:main"
-datacube-ows-update = "datacube_ows.update_ranges_impl:main"
+datacube-ows-update = "datacube_ows.update_ranges:main"
 
 [project.entry-points."datacube_ows.plugins.index"]
 postgis = "datacube_ows.index.postgis.api:ows_index_driver_init"
