+++ tests/test_update_ranges.py
@@ -12,7 +12,7 @@ import pytest
 from click.testing import CliRunner
 
 from datacube_ows.index.sql import run_sql
-from datacube_ows.update_ranges_impl import main
+from datacube_ows.update_ranges import main
 
 
 @pytest.fixture
