--- update_ranges.py
@@ -1,10 +1,0 @@
-# This file is part of datacube-ows, part of the Open Data Cube project.
-# See https://opendatacube.org for more information.
-#
-# Copyright (c) 2017-2024 OWS Contributors
-# SPDX-License-Identifier: Apache-2.0
-
-from datacube_ows.update_ranges_impl import main
-
-if __name__ == "__main__":
-    main()
